Moves header files to src/beauty
This is a common practice in c++ libraries to avoid conflicts among
imported projects. E.g. now its required to use "#include
<beauty/server.hpp>" instead of just "#include <server.hpp>"
