+++ examples/esp32/my_file_api.cpp
@@ -2,10 +2,10 @@
 #include <functional>
 #include <FS.h>
 #include <LittleFS.h>
+#include <beauty/http_result.hpp>
+#include <beauty/mime_types.hpp>
 
 #include "my_file_api.hpp"
-#include "http_result.hpp"
-#include "mime_types.hpp"
 
 using namespace http::server;
