+++ examples/pc/file_io.hpp
@@ -1,8 +1,8 @@
 #pragma once
+
 #include <fstream>
 #include <unordered_map>
-
-#include "i_file_io.hpp"
+#include <beauty/i_file_io.hpp>
 
 class FileIO : public beauty::IFileIO {
    public:
