+++ examples/pc/my_file_api.hpp
@@ -1,9 +1,8 @@
 #pragma once
 
 #include <string>
-
-#include "reply.hpp"
-#include "request.hpp"
+#include <beauty/reply.hpp>
+#include <beauty/request.hpp>
 
 class MyFileApi {
    public:
