+++ src/beauty/beauty_common.hpp
@@ -4,8 +4,8 @@
 #include <functional>
 #include <string>
 
-#include "reply.hpp"
-#include "request.hpp"
+#include "beauty/reply.hpp"
+#include "beauty/request.hpp"
 
 namespace beauty {
