+++ src/beauty/i_file_io.hpp
@@ -2,8 +2,8 @@
 
 #include <string>
 
-#include "reply.hpp"
-#include "request.hpp"
+#include "beauty/reply.hpp"
+#include "beauty/request.hpp"
 
 namespace beauty {
