+++ src/beauty/request.hpp
@@ -5,7 +5,7 @@
 #include <utility>
 #include <vector>
 
-#include "header.hpp"
+#include "beauty/header.hpp"
 
 namespace beauty {
