+++ src/connection_manager.cpp
@@ -1,4 +1,4 @@
-#include "connection_manager.hpp"
+#include "beauty/connection_manager.hpp"
 #include <chrono>
 
 namespace {
